practice-run-demo-2/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
elasticsearch==8.10.0
python-dotenv==1.0.0
